mailalerts : reset functionality added + norm

+++ mailalerts/mailalerts.php
@@ -46,12 +46,12 @@ class MailAlerts extends Module
 	{
 		$this->name = 'mailalerts';
 		$this->tab = 'administration';
-		$this->version = '3.2.2';
+		$this->version = '3.2.3';
 		$this->author = 'PrestaShop';
 		$this->need_instance = 0;
 
 		$this->bootstrap = true;
-		parent::__construct();	
+		parent::__construct();
 
 		if ($this->id)
 			$this->init();
@@ -71,7 +71,7 @@ class MailAlerts extends Module
 		$this->_product_coverage = (int)Configuration::getGlobalValue('MA_PRODUCT_COVERAGE');
 	}
 
-	public function install()
+	public function install($delete_params = true)
 	{
 		if (!parent::install() ||
 			!$this->registerHook('actionValidateOrder') ||
@@ -83,18 +83,21 @@ class MailAlerts extends Module
 			!$this->registerHook('actionProductAttributeDelete') ||
 			!$this->registerHook('actionProductAttributeUpdate') ||
 			!$this->registerHook('actionProductCoverage') ||
-			!$this->registerHook('displayHeader'))
+			!$this->registerHook('displayHeader')
+		)
 			return false;
 
-		Configuration::updateValue('MA_MERCHANT_ORDER', 1);
-		Configuration::updateValue('MA_MERCHANT_OOS', 1);
-		Configuration::updateValue('MA_CUSTOMER_QTY', 1);
-		Configuration::updateValue('MA_MERCHANT_MAILS', Configuration::get('PS_SHOP_EMAIL'));
-		Configuration::updateValue('MA_LAST_QTIES', (int)Configuration::get('PS_LAST_QTIES'));
-		Configuration::updateGlobalValue('MA_MERCHANT_COVERAGE', 0);
-		Configuration::updateGlobalValue('MA_PRODUCT_COVERAGE', 0);
+		if ($delete_params)
+		{
+			Configuration::updateValue('MA_MERCHANT_ORDER', 1);
+			Configuration::updateValue('MA_MERCHANT_OOS', 1);
+			Configuration::updateValue('MA_CUSTOMER_QTY', 1);
+			Configuration::updateValue('MA_MERCHANT_MAILS', Configuration::get('PS_SHOP_EMAIL'));
+			Configuration::updateValue('MA_LAST_QTIES', (int)Configuration::get('PS_LAST_QTIES'));
+			Configuration::updateGlobalValue('MA_MERCHANT_COVERAGE', 0);
+			Configuration::updateGlobalValue('MA_PRODUCT_COVERAGE', 0);
 
-		$sql = 'CREATE TABLE IF NOT EXISTS `'._DB_PREFIX_.MailAlert::$definition['table'].'`
+			$sql = 'CREATE TABLE IF NOT EXISTS `'._DB_PREFIX_.MailAlert::$definition['table'].'`
 				(
 					`id_customer` int(10) unsigned NOT NULL,
 					`customer_email` varchar(128) NOT NULL,
@@ -105,32 +108,46 @@ class MailAlerts extends Module
 					PRIMARY KEY  (`id_customer`,`customer_email`,`id_product`,`id_product_attribute`,`id_shop`)
 				) ENGINE='._MYSQL_ENGINE_.' DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci';
 
-		if (!Db::getInstance()->execute($sql))
+			if (!Db::getInstance()->execute($sql))
+				return false;
+		}
+
+		return true;
+	}
+
+	public function uninstall($delete_params = true)
+	{
+		if ($delete_params)
+		{
+			Configuration::deleteByName('MA_MERCHANT_ORDER');
+			Configuration::deleteByName('MA_MERCHANT_OOS');
+			Configuration::deleteByName('MA_CUSTOMER_QTY');
+			Configuration::deleteByName('MA_MERCHANT_MAILS');
+			Configuration::deleteByName('MA_LAST_QTIES');
+			Configuration::deleteByName('MA_MERCHANT_COVERAGE');
+			Configuration::deleteByName('MA_PRODUCT_COVERAGE');
+
+			if (!Db::getInstance()->execute('DROP TABLE IF EXISTS '._DB_PREFIX_.MailAlert::$definition['table']))
+				return false;
+		}
+
+		return parent::uninstall();
+	}
+
+	public function reset()
+	{
+		if (!$this->uninstall(false))
+			return false;
+		if (!$this->install(false))
 			return false;
 
 		return true;
 	}
 
-	public function uninstall()
-	{
-		Configuration::deleteByName('MA_MERCHANT_ORDER');
-		Configuration::deleteByName('MA_MERCHANT_OOS');
-		Configuration::deleteByName('MA_CUSTOMER_QTY');
-		Configuration::deleteByName('MA_MERCHANT_MAILS');
-		Configuration::deleteByName('MA_LAST_QTIES');
-		Configuration::deleteByName('MA_MERCHANT_COVERAGE');
-		Configuration::deleteByName('MA_PRODUCT_COVERAGE');
-
-		if (!Db::getInstance()->execute('DROP TABLE IF EXISTS '._DB_PREFIX_.MailAlert::$definition['table']))
-			return false;
-
-		return parent::uninstall();
-	}
-
 	public function getContent()
 	{
 		$this->_html = '';
-		
+
 		$this->_postProcess();
 
 		$this->_html .= $this->renderForm();
@@ -197,16 +214,20 @@ class MailAlerts extends Module
 
 	public function getAllMessages($id)
 	{
-		$messages = Db::getInstance()->executeS('
-			SELECT `message`
-			FROM `'._DB_PREFIX_.'message`
+		$messages = Db::getInstance()->executeS(
+			'
+						SELECT `message`
+						FROM `'._DB_PREFIX_.'message`
 			WHERE `id_order` = '.(int)$id.'
 			ORDER BY `id_message` ASC
-		');
+		'
+		);
 		$result = array();
-		foreach ($messages as $message) {
+		foreach ($messages as $message)
+		{
 			$result[] = $message['message'];
 		}
+
 		return implode('<br/>', $result);
 	}
 
@@ -222,7 +243,17 @@ class MailAlerts extends Module
 		$currency = $params['currency'];
 		$order = $params['order'];
 		$customer = $params['customer'];
-		$configuration = Configuration::getMultiple(array('PS_SHOP_EMAIL', 'PS_MAIL_METHOD', 'PS_MAIL_SERVER', 'PS_MAIL_USER', 'PS_MAIL_PASSWD', 'PS_SHOP_NAME', 'PS_MAIL_COLOR'), $id_lang, null, $id_shop);
+		$configuration = Configuration::getMultiple(
+			array(
+				'PS_SHOP_EMAIL',
+				'PS_MAIL_METHOD',
+				'PS_MAIL_SERVER',
+				'PS_MAIL_USER',
+				'PS_MAIL_PASSWD',
+				'PS_SHOP_NAME',
+				'PS_MAIL_COLOR'
+			), $id_lang, null, $id_shop
+		);
 		$delivery = new Address((int)$order->id_address_delivery);
 		$invoice = new Address((int)$order->id_address_invoice);
 		$order_date_text = Tools::displayDate($order->date_add);
@@ -266,10 +297,10 @@ class MailAlerts extends Module
 					<td style="padding:0.6em 0.4em;">'.$product['product_reference'].'</td>
 					<td style="padding:0.6em 0.4em;">
 						<strong>'
-							.$product['product_name'].(isset($product['attributes_small']) ? ' '.$product['attributes_small'] : '').(!empty($customization_text) ? '<br />'.$customization_text : '').
-						'</strong>
-					</td>
-					<td style="padding:0.6em 0.4em; text-align:right;">'.Tools::displayPrice($unit_price, $currency, false).'</td>
+				.$product['product_name'].(isset($product['attributes_small']) ? ' '.$product['attributes_small'] : '').(!empty($customization_text) ? '<br />'.$customization_text : '').
+				'</strong>
+			</td>
+			<td style="padding:0.6em 0.4em; text-align:right;">'.Tools::displayPrice($unit_price, $currency, false).'</td>
 					<td style="padding:0.6em 0.4em; text-align:center;">'.(int)$product['product_quantity'].'</td>
 					<td style="padding:0.6em 0.4em; text-align:right;">'.Tools::displayPrice(($unit_price * $product['product_quantity']), $currency, false).'</td>
 				</tr>';
@@ -277,8 +308,8 @@ class MailAlerts extends Module
 		foreach ($params['order']->getCartRules() as $discount)
 		{
 			$items_table .=
-			'<tr style="background-color:#EBECEE;">
-					<td colspan="4" style="padding:0.6em 0.4em; text-align:right;">'.$this->l('Voucher code:').' '.$discount['name'].'</td>
+				'<tr style="background-color:#EBECEE;">
+						<td colspan="4" style="padding:0.6em 0.4em; text-align:right;">'.$this->l('Voucher code:').' '.$discount['name'].'</td>
 					<td style="padding:0.6em 0.4em; text-align:right;">-'.Tools::displayPrice($discount['value'], $currency, false).'</td>
 			</tr>';
 		}
@@ -294,12 +325,18 @@ class MailAlerts extends Module
 			'{email}' => $customer->email,
 			'{delivery_block_txt}' => MailAlert::getFormatedAddress($delivery, "\n"),
 			'{invoice_block_txt}' => MailAlert::getFormatedAddress($invoice, "\n"),
-			'{delivery_block_html}' => MailAlert::getFormatedAddress($delivery, '<br />', array(
-			'firstname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>',
-			'lastname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>')),
-			'{invoice_block_html}' => MailAlert::getFormatedAddress($invoice, '<br />', array(
-			'firstname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].' font-weight:bold;">%s</span>',
-			'lastname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>')),
+			'{delivery_block_html}' => MailAlert::getFormatedAddress(
+					$delivery, '<br />', array(
+						'firstname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>',
+						'lastname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>'
+					)
+				),
+			'{invoice_block_html}' => MailAlert::getFormatedAddress(
+					$invoice, '<br />', array(
+						'firstname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].' font-weight:bold;">%s</span>',
+						'lastname' => '<span style="color:'.$configuration['PS_MAIL_COLOR'].'; font-weight:bold;">%s</span>'
+					)
+				),
 			'{delivery_company}' => $delivery->company,
 			'{delivery_firstname}' => $delivery->firstname,
 			'{delivery_lastname}' => $delivery->lastname,
@@ -341,14 +378,16 @@ class MailAlerts extends Module
 		$iso = Language::getIsoById($id_lang);
 		$dir_mail = false;
 		if (file_exists(dirname(__FILE__).'/mails/'.$iso.'/new_order.txt') &&
-		file_exists(dirname(__FILE__).'/mails/'.$iso.'/new_order.html'))
+			file_exists(dirname(__FILE__).'/mails/'.$iso.'/new_order.html')
+		)
 			$dir_mail = dirname(__FILE__).'/mails/';
 		// Send 1 email by merchant mail, because Mail::Send doesn't work with an array of recipients
 		$merchant_mails = explode(self::__MA_MAIL_DELIMITOR__, $this->_merchant_mails);
 
 		if (file_exists(_PS_MAIL_DIR_.$iso.'/new_order.txt') &&
-			file_exists(_PS_MAIL_DIR_.$iso.'/new_order.html'))
-				$dir_mail = _PS_MAIL_DIR_;
+			file_exists(_PS_MAIL_DIR_.$iso.'/new_order.html')
+		)
+			$dir_mail = _PS_MAIL_DIR_;
 
 		if ($dir_mail)
 			foreach ($merchant_mails as $merchant_mail)
@@ -386,10 +425,12 @@ class MailAlerts extends Module
 		elseif (MailAlert::customerHasNotification($id_customer, $id_product, $id_product_attribute, (int)$context->shop->id))
 			return;
 
-		$this->context->smarty->assign(array(
-			'id_product' => $id_product,
-			'id_product_attribute' => $id_product_attribute
-		));
+		$this->context->smarty->assign(
+			array(
+				'id_product' => $id_product,
+				'id_product_attribute' => $id_product_attribute
+			)
+		);
 
 		return $this->display(__FILE__, 'product.tpl');
 	}
@@ -405,11 +446,18 @@ class MailAlerts extends Module
 		$id_lang = (int)$context->language->id;
 		$product = new Product($id_product, true, $id_lang, $id_shop, $context);
 		$product_has_attributes = $product->hasAttributes();
-		$configuration = Configuration::getMultiple(array('MA_LAST_QTIES', 'PS_STOCK_MANAGEMENT', 'PS_SHOP_EMAIL', 'PS_SHOP_NAME'), null, null, $id_shop);
+		$configuration = Configuration::getMultiple(
+			array(
+				'MA_LAST_QTIES',
+				'PS_STOCK_MANAGEMENT',
+				'PS_SHOP_EMAIL',
+				'PS_SHOP_NAME'
+			), null, null, $id_shop
+		);
 		$ma_last_qties = (int)$configuration['MA_LAST_QTIES'];
 
 		$check_oos = ($product_has_attributes && $id_product_attribute) || (!$product_has_attributes && !$id_product_attribute);
-		
+
 		if ($check_oos && $product->active == 1 && (int)$quantity <= $ma_last_qties && !(!$this->_merchant_oos || empty($this->_merchant_mails)) && $configuration['PS_STOCK_MANAGEMENT'])
 		{
 			$iso = Language::getIsoById($id_lang);
@@ -421,7 +469,8 @@ class MailAlerts extends Module
 			);
 
 			if (file_exists(dirname(__FILE__).'/mails/'.$iso.'/productoutofstock.txt') &&
-				file_exists(dirname(__FILE__).'/mails/'.$iso.'/productoutofstock.html'))
+				file_exists(dirname(__FILE__).'/mails/'.$iso.'/productoutofstock.html')
+			)
 			{
 				// Send 1 email by merchant mail, because Mail::Send doesn't work with an array of recipients
 				$merchant_mails = explode(self::__MA_MAIL_DELIMITOR__, $this->_merchant_mails);
@@ -529,7 +578,8 @@ class MailAlerts extends Module
 		// if we need to send a notification
 		if ($product->active == 1 &&
 			($coverage < $warning_coverage) && !empty($this->_merchant_mails) &&
-			Configuration::getGlobalValue('MA_MERCHANT_COVERAGE'))
+			Configuration::getGlobalValue('MA_MERCHANT_COVERAGE')
+		)
 		{
 			$context = Context::getContext();
 			$id_lang = (int)$context->language->id;
@@ -543,7 +593,8 @@ class MailAlerts extends Module
 			);
 
 			if (file_exists(dirname(__FILE__).'/mails/'.$iso.'/productcoverage.txt') &&
-				file_exists(dirname(__FILE__).'/mails/'.$iso.'/productcoverage.html'))
+				file_exists(dirname(__FILE__).'/mails/'.$iso.'/productcoverage.html')
+			)
 			{
 				// Send 1 email by merchant mail, because Mail::Send doesn't work with an array of recipients
 				$merchant_mails = explode(self::__MA_MAIL_DELIMITOR__, $this->_merchant_mails);
@@ -578,7 +629,7 @@ class MailAlerts extends Module
 			$this->context->controller->addCSS($this->_path.'mailalerts.css', 'all');
 		}
 	}
-	
+
 	public function renderForm()
 	{
 		$fields_form_1 = array(
@@ -595,27 +646,27 @@ class MailAlerts extends Module
 						'name' => 'MA_CUSTOMER_QTY',
 						'desc' => $this->l('Gives the customer the option of receiving a notification for an available product if this one is out of stock.'),
 						'values' => array(
-									array(
-										'id' => 'active_on',
-										'value' => 1,
-										'label' => $this->l('Enabled')
-									),
-									array(
-										'id' => 'active_off',
-										'value' => 0,
-										'label' => $this->l('Disabled')
-									)
-								),
+							array(
+								'id' => 'active_on',
+								'value' => 1,
+								'label' => $this->l('Enabled')
+							),
+							array(
+								'id' => 'active_off',
+								'value' => 0,
+								'label' => $this->l('Disabled')
+							)
+						),
 					),
 				),
-			'submit' => array(
-				'title' => $this->l('Save'),
-				'class' => 'btn btn-default pull-right',
-				'name' => 'submitMailAlert',
+				'submit' => array(
+					'title' => $this->l('Save'),
+					'class' => 'btn btn-default pull-right',
+					'name' => 'submitMailAlert',
 				)
 			),
 		);
-		
+
 		$fields_form_2 = array(
 			'form' => array(
 				'legend' => array(
@@ -630,17 +681,17 @@ class MailAlerts extends Module
 						'name' => 'MA_MERCHANT_ORDER',
 						'desc' => $this->l('Receive a notification when an order is placed'),
 						'values' => array(
-									array(
-										'id' => 'active_on',
-										'value' => 1,
-										'label' => $this->l('Enabled')
-									),
-									array(
-										'id' => 'active_off',
-										'value' => 0,
-										'label' => $this->l('Disabled')
-									)
-								),
+							array(
+								'id' => 'active_on',
+								'value' => 1,
+								'label' => $this->l('Enabled')
+							),
+							array(
+								'id' => 'active_off',
+								'value' => 0,
+								'label' => $this->l('Disabled')
+							)
+						),
 					),
 					array(
 						'type' => 'switch',
@@ -649,17 +700,17 @@ class MailAlerts extends Module
 						'name' => 'MA_MERCHANT_OOS',
 						'desc' => $this->l('Receive a notification if the available quantity of a product is below the following threshold'),
 						'values' => array(
-									array(
-										'id' => 'active_on',
-										'value' => 1,
-										'label' => $this->l('Enabled')
-									),
-									array(
-										'id' => 'active_off',
-										'value' => 0,
-										'label' => $this->l('Disabled')
-									)
-								),
+							array(
+								'id' => 'active_on',
+								'value' => 1,
+								'label' => $this->l('Enabled')
+							),
+							array(
+								'id' => 'active_off',
+								'value' => 0,
+								'label' => $this->l('Disabled')
+							)
+						),
 					),
 					array(
 						'type' => 'text',
@@ -675,17 +726,17 @@ class MailAlerts extends Module
 						'name' => 'MA_MERCHANT_COVERAGE',
 						'desc' => $this->l('Receive a notification when an order is placed'),
 						'values' => array(
-									array(
-										'id' => 'active_on',
-										'value' => 1,
-										'label' => $this->l('Enabled')
-									),
-									array(
-										'id' => 'active_off',
-										'value' => 0,
-										'label' => $this->l('Disabled')
-									)
-								),
+							array(
+								'id' => 'active_on',
+								'value' => 1,
+								'label' => $this->l('Enabled')
+							),
+							array(
+								'id' => 'active_off',
+								'value' => 0,
+								'label' => $this->l('Disabled')
+							)
+						),
 					),
 					array(
 						'type' => 'text',
@@ -703,19 +754,18 @@ class MailAlerts extends Module
 						'desc' => $this->l('One e-mail address per line (e.g. bob@example.com)'),
 					),
 				),
-			'submit' => array(
-				'title' => $this->l('Save'),
-				'class' => 'btn btn-default pull-right',
-				'name' => 'submitMAMerchant',
+				'submit' => array(
+					'title' => $this->l('Save'),
+					'class' => 'btn btn-default pull-right',
+					'name' => 'submitMAMerchant',
 				)
 			),
 		);
 
 
-
 		$helper = new HelperForm();
 		$helper->show_toolbar = false;
-		$helper->table =  $this->table;
+		$helper->table = $this->table;
 		$lang = new Language((int)Configuration::get('PS_LANG_DEFAULT'));
 		$helper->default_form_language = $lang->id;
 		$helper->module = $this;
